ref: move verification model into module
